validate clone user input fields
