+++ app/src/main/java/com/manichord/mgit/ViewHelper.kt
@@ -1,0 +1,10 @@
+package com.manichord.mgit
+
+import android.app.Activity
+import android.content.Context
+import android.view.inputmethod.InputMethodManager
+
+fun hideKeyboard(context : Context) {
+    val inputMethodManager = context.getSystemService(Activity.INPUT_METHOD_SERVICE) as InputMethodManager
+    inputMethodManager.hideSoftInputFromWindow((context as Activity).currentFocus.windowToken, 0)
+}
