+++ app/src/main/java/com/manichord/mgit/common/BindingHelper.kt
@@ -1,0 +1,11 @@
+package com.manichord.mgit.common
+
+import android.support.design.widget.TextInputLayout
+import android.databinding.BindingAdapter
+
+
+@BindingAdapter("app:errorText")
+fun setErrorMessage(view: TextInputLayout, errorMessage: String?) {
+    errorMessage.let { view.error = errorMessage }
+}
+
